backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyMuPDF
openai
pinecone
pydantic
python-dotenv
